Conflicts:
	pom.xml

+++ src/main/java/com/bobacom/backend/dto/input/IngredienteRequest.java
@@ -1,5 +1,6 @@
 package com.bobacom.backend.dto.input;
 
+import java.math.BigDecimal;
 import java.util.List;
 
 import lombok.AllArgsConstructor;
@@ -17,9 +18,9 @@ public class IngredienteRequest {
 	private Integer id;
 	private String nome;
 	private String descrizione;
-	private Long quantitaStock;
-	private float sovraprezzoAggiunta;
-	private float prezzoRestock;
+	private BigDecimal quantitaStock;
+	private Float sovraprezzoAggiunta;
+	private Float prezzoRestock;
 	private String colore;
 	private List<Integer> idAllergene;
 	private Integer idCategoriaIngrediente;
